Search fewer overloads via static gets
Because I added get overloads for catching errors, there was more
overload searching via adl. So I made these overloads static to decrease
the search space and wrapped them with one friend get for type+index
that forwards to the overloaded static. I am not sure it's worth
it/makes a difference but it feels like the right thing to do.